refactor(vue): provide/inject и composables для экрана заказа (#196)
- ORDER_CONTEXT_KEY (Symbol), provide в OrderView, inject во вкладках с проверкой
- useOrderFormatters, useOrderFieldHelpers, useOrderLogFormatters
- Узкие props у вкладок; formatPrice передаётся в useOrderLogFormatters
- CHANGELOG: запись под 1.9.0-beta1
